Type change

+++ src/piecewiseLinearRelaxation.jl
@@ -37,7 +37,7 @@ function piecewiseLinearRelaxation!(node::SDDP.Node, plaPrecision::Dict{Symbol,A
         # Get precision
         plaPrecision_vector = Float64[]
         if nlFunction.nl_type == :emi
-            plaPrecision_vector = plaPrecision[:emi][nlIndex/2]
+            plaPrecision_vector = plaPrecision[:emi][Int(nlIndex/2)]
         elseif nlFunction.nl_type == :valve
             plaPrecision_vector = plaPrecision[:valve][Int(floor(nlIndex/2 + 1))]
         end
